Make tab_restore_service.cc compile on Mac and Linux. Review URL: http://codereview.chromium.org/23022
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9632 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/sessions/tab_restore_service.cc
@@ -10,8 +10,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/browser/browser_list.h"
 #include "chrome/browser/profile.h"
-#include "chrome/browser/sessions/session_backend.h"
 #include "chrome/browser/sessions/session_service.h"
+#include "chrome/browser/sessions/session_command.h"
 #include "chrome/browser/tab_contents/navigation_controller.h"
 #include "chrome/browser/tab_contents/navigation_entry.h"
 #include "chrome/common/scoped_vector.h"
@@ -233,7 +233,8 @@ void TabRestoreService::RestoreEntryById(Browser* browser,
         NavigationController* restored_controller =
             browser->AddRestoredTab(tab.navigations, browser->tab_count(),
                                     tab.current_navigation_index,
-                                    (tab_i == window->selected_tab_index));
+                                    (static_cast<int>(tab_i) ==
+                                     window->selected_tab_index));
         if (restored_controller)
           restored_controller->LoadIfNecessary();
       }
